Guess My Number
Guess My Number!" is a fun and interactive game built with vanilla JavaScript that runs in the browser. It's a simple yet captivating guessing game where the player tries to guess a randomly generated number between 1 and 20.

How to Play
The game starts with a randomly generated secret number that the player needs to guess.
The player enters their guess into an input field and clicks the 'Check!' button to submit their guess.
If the guess is too high or too low, the game will give a hint to help the player adjust their next guess.
The player's score decreases by 1 with each incorrect guess. The goal is to guess the number with as few attempts as possible.
If the player guesses the correct number, the game displays a winning message, and the background color changes to celebrate the victory.
The player can attempt to beat their high score in subsequent rounds.
Features
Random number generation for each game round.
Score tracking that decrements on incorrect guesses.
High score tracking across game sessions.
Interactive UI with visual cues for game status.
'Again!' button to reset the game without refreshing the page.
Technologies
HTML5 for structure.
CSS3 for styling and animations.
JavaScript for game logic and DOM manipulations.
Setup
Clone the repository, open the index.html file in your browser, and start playing!

bash
Copy code
git clone https://github.com/Sysco-CH/guess-my-number.git
Contributing
Contributions to improve the game or add new features are welcome. Please feel free to fork the repository and submit a pull request.

Acknowledgements
This game is a personal project for learning purposes, based on the teachings of Jonas Schmedtmann's JavaScript course on Udemy.